requirements.txt
Django==2.2
django-extensions==3.0.5
djangorestframework==3.11.1
ipython==7.17.0
